refactor code to align with upstream

+++ Backend/audio_backend_factory.c
@@ -27,14 +27,14 @@ static audio_backend_error_type_t map_backend_error_type(
 }
 
 capture_backend_t* audio_backend_factory_create_capture(
-    const capture_device_config_t* config, int sample_rate, int chunk_size,
-    bool full_duplex, processing_parameters_t* params,
+    const capture_device_config_t* config, size_t sample_rate,
+    size_t chunk_size, bool full_duplex, processing_parameters_t* params,
     audio_backend_error_t* out_err) {
   backend_error_t berr;
   backend_error_init(&berr, BACKEND_ERROR_NONE, "");
 
   capture_backend_t* backend = create_capture_backend(
-      config, sample_rate, chunk_size, full_duplex, params, &berr);
+      config, (int)sample_rate, (int)chunk_size, full_duplex, params, &berr);
   if (!backend || berr.type != BACKEND_ERROR_NONE) {
     logger_error(&g_logger, "Failed to create capture backend: %s",
                  berr.message);
@@ -52,14 +52,14 @@ capture_backend_t* audio_backend_factory_create_capture(
 }
 
 playback_backend_t* audio_backend_factory_create_playback(
-    const playback_device_config_t* config, int sample_rate, int chunk_size,
-    bool full_duplex, processing_parameters_t* params,
+    const playback_device_config_t* config, size_t sample_rate,
+    size_t chunk_size, bool full_duplex, processing_parameters_t* params,
     audio_backend_error_t* out_err) {
   backend_error_t berr;
   backend_error_init(&berr, BACKEND_ERROR_NONE, "");
 
   playback_backend_t* backend = create_playback_backend(
-      config, sample_rate, chunk_size, full_duplex, params, &berr);
+      config, (int)sample_rate, (int)chunk_size, full_duplex, params, &berr);
   if (!backend || berr.type != BACKEND_ERROR_NONE) {
     logger_error(&g_logger, "Failed to create playback backend: %s",
                  berr.message);
